+++ src/types.ts
@@ -44,6 +44,7 @@ export const Types = {
   Practice: Symbol('Practice'),
   IReporter: Symbol('IReporter'),
   ConfigProvider: Symbol('ConfigProvider'),
+  JSONReporter: Symbol('JSONReporter'),
 };
 
 export type GitFactory = (repository: Repository) => Git;