+++ test/helpers/mockPackage.ts
@@ -1,6 +1,6 @@
 import { DependencyType, Package } from '../../src/inspectors/IPackageInspector';
 
-export const mockPackage = (name: string, value: string = '1.0.0', dependencyType: DependencyType = DependencyType.Runtime): Package => {
+export const mockPackage = (name: string, value = '1.0.0', dependencyType: DependencyType = DependencyType.Runtime): Package => {
   const values = value.split('.');
   const major = values[0];
   const minor = values[1] || '0';